Adding path research for a monster + Start manipulating monster array. This build show a monster going from the nest to the camp following the path then get banned

+++ src/generation.c
@@ -235,8 +235,8 @@ Field generate_field() {
         }
     }
 
-    field.camp.x = (int)field.monster_path.path[field.monster_path.cur_len - 1].x;
-    field.camp.y = (int)field.monster_path.path[field.monster_path.cur_len - 1].y;
+    field.camp.x = field.monster_path.path[field.monster_path.cur_len - 1].x;
+    field.camp.y = field.monster_path.path[field.monster_path.cur_len - 1].y;
     field.board[(int)field.monster_path.path[field.monster_path.cur_len - 1].y]
                [(int)field.monster_path.path[field.monster_path.cur_len - 1].x] =
         CAMP;
